# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Inventory API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: localhost:8080
    description: Optional server description, e.g. Main (production) server
  - url: http://staging-api.example.com
    description: Optional server description, e.g. Internal staging server for testing
paths:
...
  /deleteStockById/stockId/{stockId}:
    delete:
      summary: Update stock by id.
      responses:
        '200':    # status code
          description: true should be returned
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      parameters:
        - in: path
          name: stockId
          schema:
            type: integer
          required: true
          description: Numeric ID of the stock to delete